reorganize replay dataset, full raw replay --> parser loop restored
