+++ metamon/data/replay_dataset/__init__.py
@@ -1,1 +1,1 @@
-from .loading import ShowdownReplayDataset
+
