+++ metamon/data/replay_dataset/parsed_replays/replay_parser/exceptions.py
