+++ metamon/interface.py
@@ -23,7 +23,7 @@ from poke_env.data import to_id_str
 
 import metamon
 from metamon.data import DATA_PATH
-from metamon.data.replay_dataset.replay_parser.replay_state import (
+from metamon.data.replay_dataset.parsed_replays.replay_parser.replay_state import (
     Move as ReplayMove,
     Pokemon as ReplayPokemon,
     Action as ReplayAction,
